How do I delete labels once it has been renamed?
o I have a Label where when a button is pressed, it changes the name of the text to display another label below the existing label. It pretty much just prints the same label with a different text on the row below. When the Delete Text button is pressed, it will delete the new label that has been created but how do I delete the label before the new label? I can also only remove the label via destroy and not forget.

I tried changing the text of the variable back to what it says on the window but it seems like once the name has been changed, the label doesn't exist on python but it exists on the window only.

When I enter what I want in the entry box and press Print Text, I want it to create a label. Each time I press Print Text, I want it to create a new Label beneath the old label. When I press remove item, I want it to delete the label I have created. When I keep pressing remove item, I would like it to keep removing the labels.